Simplify headers, split mut_models
Split into coati,fst,ecm,gotoh
